// src/interfaces/student.interface.ts
export interface StudentInterface {
    _id: string;
    name: string;
    studentId: string;
    dob: string;
    address: string;
    phoneNumber: string;
    prefix: string;
    typeOfSubject: string;
    FieldOfStudy: string;
    FieldOfWork: string;
    classRoom: string;
    year: string;
    
}